Add authguard. refactor todos and users to use JWT's userId

+++ server/src/users/users.controller.ts
@@ -1,11 +1,13 @@
-import { Controller, Post, Body, UsePipes, ValidationPipe, Param, Get, NotFoundException, ParseIntPipe, Delete } from '@nestjs/common';
+import { Controller, Post, Body, UsePipes, ValidationPipe, Param, Get, NotFoundException, ParseIntPipe, Delete, Request, ForbiddenException } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
+import { Public } from '../auth/public.decorator';
 
 @Controller('users')
 export class UsersController {
     constructor(private readonly usersService: UsersService) {}
 
+    @Public()
     @Post('register')
     @UsePipes(new ValidationPipe({ whitelist: true, transform: true }))
     async createUser(@Body() createUserDto: CreateUserDto) {
@@ -13,7 +15,10 @@ export class UsersController {
     }
 
     @Get(':id')
-    async findOne(@Param('id', ParseIntPipe) id: number) {
+    async findOne(@Param('id', ParseIntPipe) id: number, @Request() req) {
+        if (req.user.sub !== id) {
+            throw new ForbiddenException('You can only access your own user data');
+        }
         const user = await this.usersService.findOne(id);
         if (!user) {
             throw new NotFoundException('User not found with id: ' + id);
@@ -22,7 +27,10 @@ export class UsersController {
     }
 
     @Delete(':id')
-    async remove(@Param('id', ParseIntPipe) id: number) {
+    async remove(@Param('id', ParseIntPipe) id: number, @Request() req) {
+        if (req.user.sub !== id) {
+            throw new ForbiddenException('You can only delete your own account');
+        }
         return this.usersService.remove(id);
     }
 }
